pyproject.toml
[project]
name = "mbon-dash-2025"
version = "0.1.0"
description = "Marine Biodiversity Observatory Dashboard - Acoustic indices analysis for soundscape biodiversity assessment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs-material>=9.6.16",
    "mkdocs-mermaid2-plugin>=1.2.1",
    "mkdocs-git-revision-date-localized-plugin>=1.2.0",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "matplotlib>=3.10.5",
    "seaborn>=0.13.2",
    "marimo>=0.14.17",
    "scikit-learn>=1.7.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "requests>=2.32.4",
    "boto3>=1.40.19",
    "python-dotenv>=1.1.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mbon_analysis*", "scripts*"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-watch>=4.2.0",
]
